Add Cache-Control: public header to wrap-file-info response.
This benefits performance under a variety of conditions, as per
https://developers.google.com/speed/docs/best-practices/caching:

1. Enabling public caching in the HTTP headers for static resources allows
the browser to download resources from a nearby proxy server rather than
from a remote origin server.

2. Some versions of Firefox require that the Cache control: public header
to be set in order for resources sent over SSL to be cached on disk,
even if the other caching headers are explicitly set. Although this
header is normally used to enable caching by proxy servers (as
described below), proxies cannot cache any content sent over HTTPS,
so it is always safe to set this header for HTTPS resources.