Allow String body for wrap-params
Sometimes other middleware may have slurped the body
into a String before the wrap-params middleware gets
access to it.  As currently written, the wrap-params
middleware will still call 'slurp' on the string,
which results in trying to read a file from disk.
This commit adds a guard that will cause the middleware
to simply use the body as-is if it's already in
String format.