Support HTML5 _charset_ field in multipart params
RFC 7578 specifies that multipart text fields may use the Content-Type
header to indicate the charset used for that part. HTML5 on the other
hand, specifies that the way to set the charset is to use form field
named _charset_.

To ensure compatibility, the :encoding option still overrides the
request charset. However, a new option called :fallback-encoding now
specifies a fallback charset, which will be used if no charset can be
found.