Render stacktrace as HTML only for text/html
Changes the logic of the stacktrace middleware such that a HTML response
is returned only when the 'Accept' header contains 'text/html'. For all
other requests, a plaintext stacktrace is returned.

This maintains backward compatibility with the two previously supported
media types, while being more useful when used with non-browser clients.