Expose overriding cipher exclusions in adapter
Jetty maintains an internal default set of cipher suites and protocols
to exclude. There are use-cases for completely replacing this set rather
than adding to it. In particular, IBM JVM implementations prefix their
cipher suites with SSL_, and these are excluded by default by
Jetty's internal list.

a91d874 added the ability to extend the list of excluded cipher suites
and protocols by implementing :exclude-{ciphers,protocols}, which
provide an interface to the underlying
SslContextFactory.addExclude{CipherSuites,Protocols} methods. However,
this operation is purely additive. In order to replace the list of
excluded cipher suites, SslContextFactory.setExcludeCipherSuites needs
to be called. This commit adds this functionality.

Resolves: #404