Implement Range header middleware with bytes unit
When a Range header is given in the request, this middleware
attempts to fulfill the byte ranges specified.
If ranges are within body boundaries and not overlapping,
this middleware returns a 206 with the requested bytes in
the body. For multiple ranges, the body has a Content-Type
of multipart/byteranges with an autogenerated boundary.

The Content-Range header is set in the headers in the
case of a single given range, and in each individual part for
multiple ranges.

If any range is invalid, this middleware does nothing.

The middleware also adds a "Accept-Ranges: bytes" header to the
response.