Replace FileUpload 1.5 with FileUpload 2.0.0-M1
The Apache Commons FileUpload library has released 2.0.0-M1, the first
milestone release of 2.0.0. In this version, the core package of the
library no longer has a dependency on the Servlet API, which means we
can eliminate this problematic dependency from Ring core, once there
is a full release of FileUpload 2.0.0.