Update SPEC to fix on-open/on-close guarantees
Sometimes a HTTP request is upgraded to a WebSocket, but the WebSocket
is not correctly initiated. If the WebSocket is never successfully
opened, then on-open should not be called; similarly, on-close will
only be called if on-open has been, as if a WebSocket has not been
opened, it cannot be closed.

This commit adjusts the SPEC to specify this, updates the SPEC version
to 1.5.1, and fixes #505.