Bugfix: Make util/request-url handle empty query string
Empty query strings would make request-url insert a trailing ?
in the url for no reason. While not strictly invalid, it's a bit ugly.
Empty query strings are added by ring.mock.request/request when you pass
in an empty map for parameters.

This fixes that by checking that the query string is not empty before
adding the ?.